+++ src/api/qualicharge/api/v1/routers/statique.py
@@ -4,7 +4,7 @@ import logging
 from typing import Annotated, List
 
 from annotated_types import Len
-from fastapi import APIRouter, status
+from fastapi import APIRouter, Path, status
 from pydantic import BaseModel, computed_field
 
 from qualicharge.conf import settings
@@ -38,6 +38,39 @@ BulkStatiqueList = Annotated[List[Statique], Len(2, settings.API_BULK_CREATE_MAX
 async def list() -> List[Statique]:
     """List statique items."""
     return []
+
+
+@router.get("/{id_pdc_itinerance}")
+async def read(
+    id_pdc_itinerance: Annotated[
+        str,
+        Path(
+            description=(
+                "L'identifiant du point de recharge délivré selon les modalités "
+                "définies à l'article 10 du décret n° 2017-26 du 12 janvier 2017."
+            ),
+        ),
+    ],
+) -> Statique:
+    """Read statique item (point de charge)."""
+    raise NotImplementedError
+
+
+@router.put("/{id_pdc_itinerance}")
+async def update(
+    id_pdc_itinerance: Annotated[
+        str,
+        Path(
+            description=(
+                "L'identifiant du point de recharge délivré selon les modalités "
+                "définies à l'article 10 du décret n° 2017-26 du 12 janvier 2017."
+            ),
+        ),
+    ],
+    statique: Statique,
+) -> Statique:
+    """Update statique item (point de charge)."""
+    raise NotImplementedError
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED)