+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-04-23
+
 ### Added
 
 - Implement static data database schemas
@@ -47,7 +49,8 @@ and this project adheres to
 
 - Implement base FastAPI app
 
-[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.3.0...main
+[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.4.0...main
+[0.4.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/MTES-MCT/qualicharge/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.1.0...v0.2.0